Separate IInputType and IInputType<T>, with latter becomming IInputTypeFactory<T>
This makes it clear that IInputType doesn't have direct dependency/relation on T (and maybe some inputs can be constructed from different types of CLR objects in the future, i.e. one IInputType could have multiple factories to create it).
This addresses concerns about the T type not being used in the final input implementation (as we don't need it as it's a record of metadata for serialization)